Begins refactoring build process
Refactors codebase to centralized secrets
