+++ .bowerrc
@@ -1,0 +1,3 @@
+{
+  "directory": "dist/bower_components"
+}
