+++ bower.json
@@ -1,0 +1,9 @@
+{
+  "name": "cloudbalance",
+  "version": "0.0.1",
+  "dependencies": {
+    "jquery": "~2.1.3",
+    "oauth-js": "~0.4.0",
+    "react": "~0.12.2"
+  }
+}
