+++ package.json
@@ -22,6 +22,7 @@
     "keymirror": "^0.1.1",
     "mongoose": "~3.8.8",
     "morgan": "^1.5.1",
+    "node-env-file": "^0.1.4",
     "object-assign": "^2.0.0",
     "passport": ">= 0.0.0",
     "passport-google-oauth2": ">= 0.0.0",
@@ -34,6 +35,7 @@
     "gulp-clean": "^0.3.1",
     "gulp-concat": "^2.2.0",
     "gulp-jshint": "^1.9.0",
+    "gulp-shell": "^0.2.11",
     "react": "^0.12.2",
     "reactify": "^0.13.1"
   }
